update: refactor useFetch hook
